Fix PlaygraoundBook/build.sh

+++ PlaygroundBook/build.sh
@@ -2,6 +2,7 @@
 
 # Directories
 SELF_DIR=$(cd $(dirname $0);pwd)
+PARENT_DIR=$(cd $(dirname $0);cd ..;pwd)
 
 # Make book with template
 BOOK_DIR=$SELF_DIR/TortoisePlayground.playgroundbook
@@ -13,9 +14,12 @@ cp -rf $SELF_DIR/template/* $BOOK_DIR
 
 # Copy sources
 echo "Copy sources..."
-FROM_SOURCES_DIR=$SELF_DIR/../Sources
+FROM_SOURCES_DIR=$PARENT_DIR/Sources
 TO_SOURCES_DIR=$BOOK_DIR/Contents/Sources
-rm -f $TO_SOURCES_DIR/*
+if [ ! -e $TO_SOURCES_DIR ]; then
+  rm -f $TO_SOURCES_DIR/*
+  mkdir $TO_SOURCES_DIR
+fi
 cp $FROM_SOURCES_DIR/* $TO_SOURCES_DIR
 
 # Copy Contents.swift
